external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/extensions/webgl-debug-shaders-expected.txt

This test verifies the functionality of the WEBGL_debug_shaders extension, if it is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS WebGL context exists
PASS No WEBGL_debug_shaders support -- this is legal
PASS WEBGL_debug_shaders not listed as supported and getExtension failed -- this is legal

PASS successfullyParsed is true

TEST COMPLETE
